kafka: separate connection context from protocol
Moves connection context from being a private member of kafka::protocol
to its own top-level type. This helps keep dependencies easier to reason
about, and the split is important for organizing acl and networking
state which require different server/connection level context
information.
